It still doesnt work =[

+++ src/component/physics/MobPhysicsComponent.java
@@ -5,12 +5,12 @@ import dwarfcraft.world.World2D;
 public class MobPhysicsComponent extends PhysicsComponent {
 
 	@Override
-	public void setVelocityNorth(int speed) {
+	public void setVelocityNorth(double speed) {
 		velocity.y = speed;
 	}
 
 	@Override
-	public void setVelocityEast(int speed) {
+	public void setVelocityEast(double speed) {
 		velocity.x = speed;
 	}
 
@@ -18,8 +18,5 @@ public class MobPhysicsComponent extends PhysicsComponent {
 	public void update(World2D world, long elapsed) {
 		this.position.x += (elapsed/1000.0)*velocity.x;
 		this.position.y += (elapsed/1000.0)*velocity.y;
-		
-//		System.out.println(position);
 	}
-
 }
