+++ src/component/physics/PhysicsComponent.java
@@ -10,8 +10,8 @@ public abstract class PhysicsComponent extends Component {
 	public Point2D.Double position = new Point2D.Double();
 	public Point2D.Double velocity = new Point2D.Double();
 	
-	public abstract void setVelocityNorth(int speed);
-	public abstract void setVelocityEast(int speed);
+	public abstract void setVelocityNorth(double speed);
+	public abstract void setVelocityEast(double speed);
 	public abstract void update(World2D world, long elapsed);
 	
 	public Point2D.Double getPosition() {
